pyproject.toml
[tool.poetry]
name = "shanyrak"
version = "0.1.0"
description = ""
authors = ["dimalbek <dimalbek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
uvicorn = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.7.0"}
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
attrs = "^23.2.0"
pydantic-extra-types = "^2.6.0"
phonenumbers = "^8.13.34"
python-multipart = "^0.0.9"
datetime = "^5.5"
load-dotenv = "^0.1.0"
bcrypt = "^4.3.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
